+++ sdk/python/spacetime_memory/__init__.py
@@ -1,6 +1,7 @@
-from .client import Client
+from .client import Client, configure_logging, EmbedderUnavailableError
 from . import sdks
 from .agent_orchestrator import AgentOrchestrator
 from .metrics import MetricsCollector
 
-__all__ = ["Client", "sdks", "AgentOrchestrator", "MetricsCollector"]
+__all__ = ["Client", "sdks", "AgentOrchestrator", "MetricsCollector",
+           "configure_logging", "EmbedderUnavailableError"]